backend/requirements.txt
Django>=4.2
djangorestframework
gunicorn
psycopg2
django-environ
drf-spectacular
Pillow
numpy